+++ ndsampler/coco_dataset.py
@@ -105,6 +105,11 @@ Dataset Spec:
         We also have a new top-level dictionary to specify all the possible
         keypoint categories.
 
+Notes:
+    The main object in this file is `class`:CocoDataset, which is composed of
+    several mixin classes. See the class and method documentation for more
+    details.
+
 References:
     .. [1] http://cocodataset.org/#format-data
     .. [2] https://github.com/nightrome/cocostuffapi/blob/master/PythonAPI/pycocotools/mask.py
@@ -684,11 +689,18 @@ class MixinCocoExtras(object):
             np.ndarray : the image
         """
         import kwimage
-        gpath = self.load_image_fpath(gid_or_img)
+        gpath = self.get_image_fpath(gid_or_img)
         np_img = kwimage.imread(gpath)
         return np_img
 
     def load_image_fpath(self, gid_or_img):
+        import warnings
+        warnings.warn(
+            'get_image_fpath is deprecated use get_image_fpath instead',
+            DeprecationWarning)
+        return self.get_image_fpath(gid_or_img)
+
+    def get_image_fpath(self, gid_or_img):
         """
         Returns the full path to the image
 
@@ -934,22 +946,48 @@ class MixinCocoExtras(object):
         else:
             self.hashid_parts = None
 
-    def _ensure_imgsize(self, verbose=1):
+    def _ensure_imgsize(self, workers=0, verbose=1, fail=False):
         """
         Populate the imgsize field if it does not exist.
 
+        Args:
+            workers (int, default=0): number of workers for parallel
+                processing.
+
+            verbose (int, default=1): verbosity level
+
+            fail (bool, default=False): if True, raises an exception if
+               anything size fails to load.
+
+        Returns:
+            List[dict]: a list of "bad" image dictionaries where the size could
+            not be determined. Typically these are corrupted images and should
+            be removed.
+
         Example:
+            >>> # Normal case
             >>> self = CocoDataset.demo()
-            >>> self._ensure_imgsize()
+            >>> bad_imgs = self._ensure_imgsize()
+            >>> assert len(bad_imgs) == 0
             >>> assert self.imgs[1]['width'] == 512
             >>> assert self.imgs[2]['width'] == 300
             >>> assert self.imgs[3]['width'] == 256
+
+            >>> # Fail cases
+            >>> self = CocoDataset()
+            >>> self.add_image('does-not-exist.jpg')
+            >>> bad_imgs = self._ensure_imgsize()
+            >>> assert len(bad_imgs) == 1
+            >>> import pytest
+            >>> with pytest.raises(Exception):
+            >>>     self._ensure_imgsize(fail=True)
         """
+        bad_images = []
         if any('width' not in img or 'height' not in img
                for img in self.dataset['images']):
             from PIL import Image
 
-            def _find_imgsize(gpath):
+            def _find_imgshape(gpath):
                 try:
                     pil_img = Image.open(gpath)
                     w, h = pil_img.size
@@ -960,21 +998,37 @@ class MixinCocoExtras(object):
                         dset = gdal.Open(gpath, gdal.GA_ReadOnly)
                         w = dset.RasterXSize
                         h = dset.RasterYSize
-                    except ImportError:
+                    except Exception:
                         raise pil_ex
-                return w, h
+                return h, w
+            # TODO: use kwimage.im_io.load_image_shape instead
 
             if self.tag:
                 desc = 'populate imgsize for ' + self.tag
             else:
                 desc = 'populate imgsize for untagged coco dataset'
-            for img in ub.ProgIter(self.dataset['images'], desc=desc,
-                                   verbose=verbose):
+
+            from ndsampler import util_futures
+            pool = util_futures.JobPool('thread', max_workers=workers)
+            for img in ub.ProgIter(self.dataset['images'], verbose=verbose,
+                                   desc='submit image size jobs'):
                 gpath = join(self.img_root, img['file_name'])
-                if 'width' not in img:
-                    w, h = _find_imgsize(gpath)
-                    img['width'] = w
-                    img['height'] = h
+                if 'width' not in img or 'height' not in img:
+                    job = pool.submit(_find_imgshape, gpath)
+                    job.img = img
+
+            for job in ub.ProgIter(pool.as_completed(), total=len(pool),
+                                   verbose=verbose, desc=desc):
+                try:
+                    h, w = job.result()[0:2]
+                except Exception:
+                    if fail:
+                        raise
+                    bad_images.append(job.img)
+                else:
+                    job.img['width'] = w
+                    job.img['height'] = h
+        return bad_images
 
     def _resolve_to_id(self, id_or_dict):
         """
@@ -1967,13 +2021,13 @@ class MixinCocoDraw(object):
             else:
                 raise Exception('no bbox, line, or keypoint position')
 
-            cid = ann['category_id']
+            cid = ann.get('category_id', None)
             if cid is not None:
                 cat = self.cats[cid]
                 catname = cat['name']
             else:
                 cat = None
-                catname = 'None'
+                catname = ann.get('category_name', 'None')
             textkw = {
                 'horizontalalignment': 'left',
                 'verticalalignment': 'top',
@@ -2273,6 +2327,9 @@ class MixinCocoAddRemove(object):
         """
         Like add_category, but returns the existing category id if it already
         exists instead of failing. In this case all metadata is ignored.
+
+        Returns:
+            int: the existing or new category id
         """
         try:
             id = self.add_category(name=name, supercategory=supercategory,
@@ -3325,15 +3382,20 @@ class CocoDataset(ub.NiceRepr, MixinCocoAddRemove, MixinCocoStats,
     def _clear_index(self):
         self.index.clear()
 
-    def union(self, *others, **kw):
+    def union(self, *others, **kwargs):
         """
         Merges multiple `CocoDataset` items into one. Names and associations
         are retained, but ids may be different.
 
-        TODO: are supercategories broken?
+        Args:
+            self : note that `union` can be called as an instance method or a class method.
+                If it is a class method, then this is the class type, otherwise the instance
+                will also be unioned with `others`.
+            *others : a series of CocoDatasets that we will merge
+            **kwargs : constructor options for the new merged CocoDataset
 
-        CommandLine:
-            xdoctest -m ~/code/ndsampler/ndsampler/coco_dataset.py CocoDataset.union
+        Returns:
+            CocoDataset: a new merged coco dataset
 
         Example:
             >>> # Test union works with different keypoint categories
@@ -3354,10 +3416,43 @@ class CocoDataset(ub.NiceRepr, MixinCocoAddRemove, MixinCocoStats,
             >>> assert kpfreq_want == kpfreq_got1
             >>> assert kpfreq_want == kpfreq_got2
 
-        Ignore:
-            dset_12.dataset['keypoint_categories']
-            dset_12._keypoint_category_names()
-            dset_21._keypoint_category_names()
+            >>> # Test disjoint gid datasets
+            >>> import ndsampler
+            >>> dset1 = ndsampler.CocoDataset.demo('shapes3')
+            >>> for new_gid, img in enumerate(dset1.dataset['images'], start=10):
+            >>>     for aid in dset1.gid_to_aids[img['id']]:
+            >>>         dset1.anns[aid]['image_id'] = new_gid
+            >>>     img['id'] = new_gid
+            >>> dset1._clear_index()
+            >>> dset1._build_index()
+            >>> # ------
+            >>> dset2 = ndsampler.CocoDataset.demo('shapes2')
+            >>> for new_gid, img in enumerate(dset2.dataset['images'], start=100):
+            >>>     for aid in dset2.gid_to_aids[img['id']]:
+            >>>         dset2.anns[aid]['image_id'] = new_gid
+            >>>     img['id'] = new_gid
+            >>> dset2._clear_index()
+            >>> dset2._build_index()
+            >>> others = [dset1, dset2]
+            >>> merged = ndsampler.CocoDataset.union(*others)
+            >>> print('merged = {!r}'.format(merged))
+            >>> print('merged.imgs = {}'.format(ub.repr2(merged.imgs, nl=1)))
+            >>> assert set(merged.imgs) & set([10, 11, 12, 100, 101]) == set(merged.imgs)
+
+            >>> # Test data is not preserved
+            >>> dset2 = ndsampler.CocoDataset.demo('shapes2')
+            >>> dset1 = ndsampler.CocoDataset.demo('shapes3')
+            >>> others = (dset1, dset2)
+            >>> cls = self = ndsampler.CocoDataset
+            >>> merged = cls.union(*others)
+            >>> print('merged = {!r}'.format(merged))
+            >>> print('merged.imgs = {}'.format(ub.repr2(merged.imgs, nl=1)))
+            >>> assert set(merged.imgs) & set([1, 2, 3, 4, 5]) == set(merged.imgs)
+
+        TODO:
+            - [ ] are supercategories broken?
+            - [ ] reuse image ids where possible
+            - [ ] reuse annotation / category ids where possible
         """
         if self.__class__ is type:
             # Method called as classmethod
@@ -3388,6 +3483,18 @@ class CocoDataset(ub.NiceRepr, MixinCocoAddRemove, MixinCocoStats,
                     if k not in d1:
                         d1[k] = v
                 return d1
+
+            def _has_duplicates(items):
+                seen = set()
+                for item in items:
+                    if item in seen:
+                        return True
+                    seen.add(item)
+                return False
+
+            _all_imgs = (img for _, d in relative_dsets for img in d['images'])
+            _all_gids = (img['id'] for img in _all_imgs)
+            preserve_gids = not _has_duplicates(_all_gids)
 
             for subdir, old_dset in relative_dsets:
                 # Create temporary indexes to map from old to new
@@ -3456,8 +3563,12 @@ class CocoDataset(ub.NiceRepr, MixinCocoAddRemove, MixinCocoStats,
 
                 # Add the images into the merged dataset
                 for old_img in old_dset['images']:
+                    if preserve_gids:
+                        new_id = old_img['id']
+                    else:
+                        new_id = len(merged['images']) + 1
                     new_img = _dict([
-                        ('id', len(merged['images']) + 1),
+                        ('id', new_id),
                         ('file_name', join(subdir, old_img['file_name'])),
                     ])
                     # copy over other metadata
@@ -3517,7 +3628,7 @@ class CocoDataset(ub.NiceRepr, MixinCocoAddRemove, MixinCocoStats,
         if common_root is not None:
             merged['img_root'] = common_root
 
-        new_dset = cls(merged, **kw)
+        new_dset = cls(merged, **kwargs)
         return new_dset
 
     def subset(self, gids, copy=False):